Pull out s3 code as subcommand
